+++ src/refs.ts
@@ -1,0 +1,5 @@
+import { RefObject } from './types';
+
+export const createRef = <TRef>(): RefObject<TRef> => ({
+  current: null,
+});
